MAJOR uwu.py refactors and splitting
