+++ setup.py
@@ -10,7 +10,6 @@
 # the Free Software Foundation, either version 3 of the License, or
 # (at your option) any later version.
 
-import os
 import sys
 import subprocess
 import tomllib
