+++ utils/constants.py
@@ -1,0 +1,13 @@
+from rich.panel import Panel
+from rich.align import Align
+owo_dusk_api = "https://echoquill.github.io/owo-dusk-api"
+
+owoArt = r"""
+  __   _  _   __       ____  _  _  ____  __ _ 
+ /  \ / )( \ /  \  ___(    \/ )( \/ ___)(  / )
+(  O )\ /\ /(  O )(___)) D () \/ (\___ \ )  ( 
+ \__/ (_/\_) \__/     (____/\____/(____/(__\_)
+"""
+owoPanel = Panel(Align.center(owoArt), style="purple ", highlight=False)
+version = "2.5.0"
+database_version = "2.5.0"
