+++ utils/state.py
@@ -1,0 +1,2 @@
+# do not delete, made to avoid circular imports between uwu and utils.popup
+popup_queue = None
